requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
coverage==4.2
coveralls==1.1
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
django-cors-headers==1.1.0
django-dual-authentication==1.0.0
django-envie==0.0.9
django-filter==0.13.0
django-nose==1.4.3
django-registration-redux==1.4
django-rest-swagger==0.3.6
django-toolbelt==0.0.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
docopt==0.6.2
drf-nested-routers==0.11.1
drfdocs==0.0.11
fake-factory==0.5.9
futures==3.0.5
gunicorn==19.3.0
ipaddress==1.0.16
ipython==5.0.0
ipython-genutils==0.1.0
Markdown==2.6.6
nose==1.3.7
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.0
python-dateutil==2.5.3
PyYAML==3.11
requests==2.10.0
requests-futures==0.9.7
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
traitlets==4.2.2
wcwidth==0.1.7
